Added some connexion feature (can now connect only with email)
Added new public route to see word cloud

Added the possibility to stop a quiz while it's currently playing

+++ src/main/java/com/pixsom/racebets/quiz/dto/QuizSessionSummaryResponse.java
@@ -9,6 +9,6 @@ public record QuizSessionSummaryResponse(
         QuizSessionPhase phase,
         int currentQuestionIndex,
         int questionCount,
-        int participantCount
+        long participantCount
 ) {
 }
